requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
argparse>=1.4.0
